+++ proof_gen/src/proof_gen.rs
@@ -4,12 +4,18 @@
 use std::sync::{atomic::AtomicBool, Arc};
 
 use evm_arithmetization::{AllStark, StarkConfig};
-use plonky2::util::timing::TimingTree;
+use plonky2::{
+    gates::noop::NoopGate,
+    iop::witness::PartialWitness,
+    plonk::{circuit_builder::CircuitBuilder, circuit_data::CircuitConfig},
+    util::timing::TimingTree,
+};
 use trace_decoder::types::TxnProofGenIR;
 
 use crate::{
     proof_types::{AggregatableProof, GeneratedAggProof, GeneratedBlockProof, GeneratedTxnProof},
     prover_state::ProverState,
+    types::{Config, Field, PlonkyProofIntern, EXTENSION_DEGREE},
 };
 
 /// A type alias for `Result<T, ProofGenError>`.
@@ -107,4 +113,22 @@ pub fn generate_block_proof(
         b_height,
         intern: b_proof_intern,
     })
+}
+
+/// Generates a dummy proof for a dummy circuit doing nothing.
+/// This is useful for testing purposes only.
+pub fn dummy_proof() -> ProofGenResult<PlonkyProofIntern> {
+    let mut builder = CircuitBuilder::<Field, EXTENSION_DEGREE>::new(CircuitConfig::default());
+    builder.add_gate(NoopGate, vec![]);
+    let circuit_data = builder.build::<_>();
+
+    let inputs = PartialWitness::new();
+
+    plonky2::plonk::prover::prove::<Field, Config, EXTENSION_DEGREE>(
+        &circuit_data.prover_only,
+        &circuit_data.common,
+        inputs,
+        &mut TimingTree::default(),
+    )
+    .map_err(|e| ProofGenError(e.to_string()))
 }